feat(storybook): add abstraction
